Add Docker-based testing setup for Spark applications
- Introduced `conftest.py` to manage Docker service setup for tests.
- Refactored test scripts to utilize the new `run_docker` function for executing Spark jobs.
- Added validation scripts for various outputs including listings, reviews, and weather data.
- Updated test cases to reflect new validation paths and ensure consistency in data validation checks.
